+++ src/core/Analytics.ts
@@ -3,12 +3,12 @@ import Constants from 'expo-constants';
 
 import appConfig from '../../appConfig';
 import UserService from './user/UserService';
-import { userService } from '@covid/Services';
 
 let isInitialized = false;
 
 type AdditionalUserProperties = {
-  isTester: boolean;
+  isTester?: boolean;
+  Experiment_001?: string;
 };
 
 export const events = {
@@ -20,10 +20,6 @@ export const events = {
   JOIN_STUDY: 'JOIN_STUDY',
   DECLINE_STUDY: 'DECLINE_STUDY',
   CLICK_CALLOUT: 'CLICK_CALLOUT',
-};
-
-export const experiments = {
-  Experiment_001: 'Experiment_001', // Test alternative external callouts on UK Thank You Pags
 };
 
 // Disable Tracking of the User Properties (Only available in Expo SDK 37)
@@ -75,42 +71,9 @@ export function identify(additionalProps?: AdditionalUserProperties): void {
   Amplitude.setUserProperties(payload);
 }
 
-function hashToInt(s: string): number {
-  // Implementation of a quick and evenly spread hash to an integer:
-  // https://stackoverflow.com/questions/7616461/generate-a-hash-from-string-in-javascript
-  let hash = 0,
-    i,
-    chr;
-  for (i = 0; i < s.length; i++) {
-    chr = s.charCodeAt(i);
-    hash = (hash << 5) - hash + chr;
-    hash |= 0; // Convert to 32bit integer
-  }
-  return Math.abs(hash);
-}
-
-function getVariant(hash: string, totalVariants: number): string {
-  const variantNumber = (hashToInt(hash) % totalVariants) + 1;
-  return 'variant_' + variantNumber;
-}
-
-export async function startExperiment(experimentName: string, totalVariants: number): Promise<string | null> {
-  const profile = await userService.getProfile();
-  if (!profile) return null;
-
-  const variant = getVariant(profile.username, totalVariants);
-  const payload: { [index: string]: string } = {};
-  payload[experimentName] = variant;
-
-  initialize();
-  Amplitude.setUserProperties(payload);
-  return variant;
-}
-
 export default {
   track,
   trackScreenView,
   events,
-  experiments,
   identify,
 };
